DX Localhost: Solved the styling issues and refactor code
Separate the styling of adminbar and tooblar button so if the toolbar is
disable user still can change the admin bar color. Refactor the code so
the get_option function should not call again and again without any
need. In that case we have reduced query. Tested with wordpress 5.0.3
version.

Asana Task: DX localhost: test the plugin with the newest version of
WordPress Core